# documentation/api.yaml
openapi: 3.0.0
info:
  title: Documentation for the ERP @ablazelabs
  description: this documentation was made by swagger-ui
  version: "1.0"
  contact: 
    email: 'yaredterefeg@gmail.com'
  license:
    name: Ablaze Labs
    url: https://www.ablazelabs.com
paths:
...
  /role:
    post:
      summary: Creates a new role
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                privileges:
                  type: array
                  items:
                    type: integer
                description:
                  type: string
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    get:
      summary: gets selected roles
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                skip:
                  type: integer
                filter:
                  type: object
                  properties:
                    name:
                      type: string
                    description:
                      type: string
                sort:
                  type: object
                  properties:
                    name:
                      type: integer
                    description:
                      type: integer
                    id:
                      type: integer
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
                    concurrency_stamp:
                      type: string
                    privileges:
                      type: array
                      items:
                        type: object
    patch:
      summary: updates a role
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                concurrency_stamp:
                  type: string
                id:
                  type: integer
                updateData:
                  type: object
                  properties:
                    name:
                      type: string
                    privileges:
                      type: array
                      items:
                        type: integer
                    description:
                      type: string
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    delete:
      summary: updates a role
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
      responses:
        200:
          description: success message
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...